Lab #3 update

+++ programming/lab3/lab/records/Eatable.java
@@ -3,8 +3,8 @@ package lab.records;
 import lab.interfaces.ICapitalisticPassive;
 import lab.interfaces.IMeasurable;
 
-public record Eatable(String name, int saturation, float cost) implements ICapitalisticPassive, IMeasurable {
-   public Eatable(String name, int saturation) {
+public record Eatable(String name, byte saturation, float cost) implements ICapitalisticPassive, IMeasurable {
+   public Eatable(String name, byte saturation) {
       this(name, saturation, ICapitalisticPassive.DEFAULT_COST);
    }
 }
